streamlit_app/requirements.txt
streamlit
pandas
azure-storage-blob
pyarrow